domain arrangement code
